refactor: 폴더구조변경 및 배경 변경
