+++ src/pages/service/components/ImageSkeleton.tsx
@@ -1,0 +1,6 @@
+export const ImageSkeleton = () => (
+  <div className='group relative h-40 animate-pulse overflow-hidden rounded-lg'>
+    <div className='h-full w-full' />
+    <div className='absolute bottom-0 left-0 right-0 h-8' />
+  </div>
+);
